ostree-prepare-root: Allow ostree= argument to be passed on command line
This is useful if running an image of an OSTree OS (an .img file) under
systemd-nspawn: use ostree-prepare-root as the initial PID1 within the
nspawn container to do the pivot root and mount setup, then hand off to
the real init system. In this case, /proc/cmdline comes from the host
system, and doesn’t reflect what’s in /ostree in the container.

Signed-off-by: Philip Withnall <withnall@endlessm.com>