ostree-prepare-root: Allow init path to be passed on command line
If using systemd-nspawn to boot an OSTree OS from an image (.img file),
ostree-prepare-root needs to be run as the initial PID1 in order to set
up the mounts and pivot root within the container. It can then exec
/sbin/init in order to continue booting the system — this essentially
implements the --boot option of systemd-nspawn.

However, the other mode for systemd-nspawn – running a single non-init
process – cannot be run this way. This commit fixes that by allowing the
argv to be exec()ed to be specified on the command line for
ostree-prepare-root. It defaults to /sbin/init and inherits the
environment from ostree-prepare-root as before.

Signed-off-by: Philip Withnall <withnall@endlessm.com>