libostree: Add Avahi based OstreeRepoFinder implementation
This is a more complex implementation of OstreeRepoFinder which resolves
ref names to remote URIs by looking for refs advertised by peers on the
local network using DNS-SD records and mDNS (Avahi). The idea is to
allow OS and app updates to be propagated over local networks, without
the internet.

It requires an OSTree server and code to generate the DNS-SD adverts in
order to be fully functional — support for this will be added
separately.

This commit is unpolished and not ready to be merged. It is intended as
a strawman for feedback.

Unit tests are included.

Signed-off-by: Philip Withnall <withnall@endlessm.com>